优化后台管理

+++ app/Http/Controllers/Api/V1/PostCommentsController.php
@@ -10,6 +10,8 @@ class PostCommentsController extends Controller
 {
     public function destroy(PostComment $comment)
     {
+        $this->authorize('own', $comment);
+
         $comment->delete();
 
         return response()->noContent();
